mock app CLI: support legacy username/password logins
If we have the app client secret, use it to (try and) fetch a token using
"Resource Owner Password Credentials Grant Type.

Because this is a playground, we just fork a process to ask for input(),
such that we can just kill it if we get a token via web app.  Something
elegant using poll() could be done instead.